+++ examples/animations/main.go
@@ -207,6 +207,7 @@ func detailView(w *gui.Window) gui.View {
 
 func animButton(label string, action func(w *gui.Window)) gui.View {
 	return gui.Button(gui.ButtonCfg{
+		ID:      "animations_anim_button",
 		Content: []gui.View{gui.Text(gui.TextCfg{Text: label})},
 		OnClick: func(ctx gui.EventCtx) {
 			action(ctx.Window)